+++ embassy-stm32/src/eth/v2/descriptors.rs
